+++ regexSearch.py
@@ -1,0 +1,94 @@
+import re
+from kwic import afficher_kwic
+
+def requete_regex(corpus):
+        
+    pattern = input("Entrez l'expression régulière : ").strip()
+
+    if not pattern:
+        print("Expression invalide, retour au menu.")
+        return
+
+    type = input("Recherche sur le mot ou sur son étiquette grammaticale ? (mot/tag, defaut mot) ").strip().lower()
+    while type not in ('mot', 'tag'):
+        type = input("Veuillez répondre 'mot' ou 'tag' : ").strip().lower()
+
+    kwic_choix = input("Afficher le contexte (KWIC) ? (oui/non, défaut non) : ").strip().lower()
+    kwic = (kwic_choix == 'oui')
+
+    if kwic:
+        size = int(input("Entrez le nombre de mots à gauche et à droite que vous souhaitez afficher, par défaut 5) : ") or 5)
+        case_choix = input("Respecter la casse ? (oui/non, défaut non) : ").strip().lower()
+        case_sensitive = (case_choix == 'oui')
+
+        results = corpus.kwic_regex(pattern, type = type, size = size, case_sensitive = case_sensitive)
+        if results:
+            corpus.afficher_kwic(results, size = size)
+        else:
+            print("Aucune occurrence trouvée.")
+
+    else:
+        case_choix = input("Respecter la casse ? (oui/non, défaut non) : ").strip().lower()
+        case_sensitive = (case_choix == 'oui')
+
+        regex = re.compile(pattern)
+        results = []
+
+        for id_phrase, phrase in enumerate(corpus.sentences, start=1):
+            for pos, token in enumerate(phrase, start=1):
+                mot, tag, token_new = corpus.parser_token(token)
+                if type == 'tag':
+                    cible = tag
+                else:
+                    if tag != 'NPP' and not case_sensitive:
+                        cible = mot.lower()
+                    else:
+                        cible = mot
+
+                if regex.search(cible):
+                    results.append({
+                        'mot': mot,
+                        'tag': tag,
+                        'phrase_id': id_phrase,
+                        'pos': pos
+                    })
+        if results:
+            print(f"\n{len(results)} occurrence(s) trouvée(s) :")
+            for res in results[:20]:
+                print(f"Phrase {res['phrase_id']}, position {res['pos']} : {res['mot']}/{res['tag']}")
+            if len(results) > 20:
+                print(f"... et {len(results)-20} autres.")
+        else:
+            print("Aucune occurrence trouvée.")
+
+
+
+def kwic_regex(corpus, pattern, type = 'mot', size = 5, case_sensitive = False):
+
+    regex = re.compile(pattern)
+    results = []
+
+    for id_phrase, phrase in enumerate(corpus.sentences, start=1):
+        for pos, token in enumerate(phrase, start=1):
+            mot, tag, token_new = corpus.parser_token(token)
+            if type == 'tag':
+                cible = tag
+            else:
+                if tag != 'NPP' and not case_sensitive:
+                    cible = mot.lower()
+                else:
+                    cible = mot
+
+            if regex.search(cible):
+                  
+                gauche_ind = max(0, pos - 1 - size) 
+                droite_ind = min(len(phrase), pos + size) 
+                results.append({
+                    'id_phrase': id_phrase,
+                    'pos': pos,
+                    'gauche': phrase[gauche_ind:pos-1],
+                    'mot_enquete': token,
+                    'droite': phrase[pos:droite_ind]
+                })
+
+    return results
